content: generateGroupInfoPage: content function syntax cleanup
