+++ src/content/dependencies/generateGroupInfoPage.js
@@ -1,221 +1,90 @@
-import {empty, stitchArrays} from '#sugar';
-
 export default {
   contentDependencies: [
-    'generateAbsoluteDatetimestamp',
-    'generateColorStyleAttribute',
-    'generateContentHeading',
+    'generateGroupInfoPageAlbumsSection',
     'generateGroupNavLinks',
     'generateGroupSecondaryNav',
     'generateGroupSidebar',
     'generatePageLayout',
-    'linkAlbum',
     'linkExternal',
-    'linkGroupGallery',
-    'linkGroup',
     'transformContent',
   ],
 
   extraDependencies: ['html', 'language', 'wikiData'],
 
-  sprawl({wikiInfo}) {
-    return {
-      enableGroupUI: wikiInfo.enableGroupUI,
-    };
-  },
+  sprawl: ({wikiInfo}) => ({
+    enableGroupUI:
+      wikiInfo.enableGroupUI,
+  }),
 
-  query(sprawl, group) {
-    const albums =
-      group.albums;
+  relations: (relation, sprawl, group) => ({
+    layout:
+      relation('generatePageLayout'),
 
-    const albumGroups =
-      albums
-        .map(album => album.groups);
+    navLinks:
+      relation('generateGroupNavLinks', group),
 
-    const albumOtherCategory =
-      albumGroups
-        .map(groups => groups
-          .map(group => group.category)
-          .find(category => category !== group.category));
+    secondaryNav:
+      (sprawl.enableGroupUI
+        ? relation('generateGroupSecondaryNav', group)
+        : null),
 
-    const albumOtherGroups =
-      stitchArrays({
-        groups: albumGroups,
-        category: albumOtherCategory,
-      }).map(({groups, category}) =>
-          groups
-            .filter(group => group.category === category));
+    sidebar:
+      (sprawl.enableGroupUI
+        ? relation('generateGroupSidebar', group)
+        : null),
 
-    return {albums, albumOtherGroups};
-  },
-
-  relations(relation, query, sprawl, group) {
-    const relations = {};
-    const sec = relations.sections = {};
-
-    relations.layout =
-      relation('generatePageLayout');
-
-    relations.navLinks =
-      relation('generateGroupNavLinks', group);
-
-    if (sprawl.enableGroupUI) {
-      relations.secondaryNav =
-        relation('generateGroupSecondaryNav', group);
-
-      relations.sidebar =
-        relation('generateGroupSidebar', group);
-    }
-
-    sec.info = {};
-
-    sec.info.visitLinks =
+    visitLinks:
       group.urls
-        .map(url => relation('linkExternal', url));
+        .map(url => relation('linkExternal', url)),
 
-    if (group.description) {
-      sec.info.description =
-        relation('transformContent', group.description);
-    }
+    description:
+      relation('transformContent', group.description),
 
-    if (!empty(query.albums)) {
-      sec.albums = {};
+    albumSection:
+      relation('generateGroupInfoPageAlbumsSection', group),
+  }),
 
-      sec.albums.heading =
-        relation('generateContentHeading');
+  data: (_sprawl, group) => ({
+    name:
+      group.name,
 
-      sec.albums.galleryLink =
-        relation('linkGroupGallery', group);
+    color:
+      group.color,
+  }),
 
-      sec.albums.albumColorStyles =
-        query.albums
-          .map(album => relation('generateColorStyleAttribute', album.color));
+  generate: (data, relations, {html, language}) =>
+    relations.layout.slots({
+      title: language.$('groupInfoPage.title', {group: data.name}),
+      headingMode: 'sticky',
+      color: data.color,
 
-      sec.albums.albumLinks =
-        query.albums
-          .map(album => relation('linkAlbum', album));
+      mainContent: [
+        html.tag('p',
+          {[html.onlyIfContent]: true},
+          language.$('releaseInfo.visitOn', {
+            [language.onlyIfOptions]: ['links'],
+            links:
+              language.formatDisjunctionList(
+                relations.visitLinks
+                  .map(link => link.slot('context', 'group'))),
+          })),
 
-      sec.albums.otherGroupLinks =
-        query.albumOtherGroups
-          .map(groups => groups
-            .map(group => relation('linkGroup', group)));
+        html.tag('blockquote',
+          {[html.onlyIfContent]: true},
+          relations.description.slot('mode', 'multiline')),
 
-      sec.albums.datetimestamps =
-        group.albums.map(album =>
-          (album.date
-            ? relation('generateAbsoluteDatetimestamp', album.date)
-            : null));
-    }
+        relations.albumSection,
+      ],
 
-    return relations;
-  },
+      leftSidebar:
+        (relations.sidebar
+          ? relations.sidebar
+              .content /* TODO: Kludge. */
+          : null),
 
-  data(query, sprawl, group) {
-    const data = {};
+      navLinkStyle: 'hierarchical',
+      navLinks: relations.navLinks.content,
 
-    data.name = group.name;
-    data.color = group.color;
-
-    return data;
-  },
-
-  generate(data, relations, {html, language}) {
-    const {sections: sec} = relations;
-
-    return relations.layout
-      .slots({
-        title: language.$('groupInfoPage.title', {group: data.name}),
-        headingMode: 'sticky',
-        color: data.color,
-
-        mainContent: [
-          html.tag('p',
-            {[html.onlyIfContent]: true},
-            language.$('releaseInfo.visitOn', {
-              [language.onlyIfOptions]: ['links'],
-              links:
-                language.formatDisjunctionList(
-                  sec.info.visitLinks
-                    .map(link => link.slot('context', 'group'))),
-            })),
-
-          html.tag('blockquote',
-            {[html.onlyIfContent]: true},
-            sec.info.description
-              ?.slot('mode', 'multiline')),
-
-          sec.albums && [
-            sec.albums.heading
-              .slots({
-                tag: 'h2',
-                title: language.$('groupInfoPage.albumList.title'),
-              }),
-
-            html.tag('p',
-              language.$('groupInfoPage.viewAlbumGallery', {
-                link:
-                  sec.albums.galleryLink
-                    .slot('content', language.$('groupInfoPage.viewAlbumGallery.link')),
-              })),
-
-            html.tag('ul',
-              stitchArrays({
-                albumLink: sec.albums.albumLinks,
-                otherGroupLinks: sec.albums.otherGroupLinks,
-                datetimestamp: sec.albums.datetimestamps,
-                albumColorStyle: sec.albums.albumColorStyles,
-              }).map(({
-                  albumLink,
-                  otherGroupLinks,
-                  datetimestamp,
-                  albumColorStyle,
-                }) => {
-                  const prefix = 'groupInfoPage.albumList.item';
-                  const parts = [prefix];
-                  const options = {};
-
-                  options.album =
-                    albumLink.slot('color', false);
-
-                  if (datetimestamp) {
-                    parts.push('withYear');
-                    options.yearAccent =
-                      language.$(prefix, 'yearAccent', {
-                        year:
-                          datetimestamp.slots({style: 'year', tooltip: true}),
-                      });
-                  }
-
-                  if (!empty(otherGroupLinks)) {
-                    parts.push('withOtherGroup');
-                    options.otherGroupAccent =
-                      html.tag('span', {class: 'other-group-accent'},
-                        language.$(prefix, 'otherGroupAccent', {
-                          groups:
-                            language.formatConjunctionList(
-                              otherGroupLinks.map(groupLink =>
-                                groupLink.slot('color', false))),
-                        }));
-                  }
-
-                  return (
-                    html.tag('li',
-                      albumColorStyle,
-                      language.$(...parts, options)));
-                })),
-          ],
-        ],
-
-        leftSidebar:
-          (relations.sidebar
-            ? relations.sidebar
-                .content /* TODO: Kludge. */
-            : null),
-
-        navLinkStyle: 'hierarchical',
-        navLinks: relations.navLinks.content,
-
-        secondaryNav: relations.secondaryNav ?? null,
-      });
-  },
+      secondaryNav: relations.secondaryNav ?? null,
+    }),
 };
